% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choleskiDecomposition.R
\name{choleskiDecomposition}
\alias{choleskiDecomposition}
\title{Function for Choleski Decomposition}
\usage{
choleskiDecomposition(A, tol = 1e-07)
}
\arguments{
\item{A}{:Input Matrix}

\item{tol}{: Tolerance}
}
\value{
L: Decomposition matrix
}
\description{
Function perform choleski decomposition for positive definate matrix (A=LL')
}
\examples{
A<-matrix(c(4,-2,2, -2,2,-4,2,-4,11), nrow=3,ncol=3, byrow = TRUE)
chdA<-choleskiDecomposition(A)
}

