% Generated by roxygen2 (4.0.1): do not edit by hand
\name{choleskilm}
\alias{choleskilm}
\title{Function fits linear model using Choleski Decomposition}
\usage{
choleskilm(A, b, tol = 1e-07)
}
\arguments{
\item{A}{: Input matrix}

\item{b}{: Response matrix}

\item{tol}{: Tolerance}
}
\value{
U  : Upper part of the triangele is (U) and Lower part of the triangular is L (Diagnoal for the L matrix is 1)

c  : Lc=b (where Ux=c)

beta : Estimates

examples
A<-matrix(c(6,-4,1, -4,6,-4,1,-4,6), nrow=3,ncol=3, byrow = TRUE)
b<-matrix(c(-14,36, 6), nrow=3,ncol=1,byrow=TRUE)
Z<-optR(A, b, method="choleski") # Solve Linear model using Gauss Elimination
}
\description{
Function fits a linear model using Choleski Decomposition for positive definate matrix
}

