% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hatMatrix}
\alias{hatMatrix}
\title{Function determines the Hat matrix or projection matrix for given X}
\usage{
hatMatrix(X, covmat = NULL)
}
\arguments{
\item{X}{: Input Matrix}

\item{covmat}{: covariance matrix for error, if the error are correlated for I.I.D covmat will be NULL matrix}
}
\value{
X: Projection Matrix
}
\description{
Function hatMatrix determines the projection matrix for X from the form yhat=Hy. The projection matrix defines the influce of each variable on fitted value
The diagonal elements of the projection matrix are the leverages or influence each sample has on the fitted value for that same observation.
The projection matrix is evaluated with I.I.D assumtion ~N(0, 1)
}
\examples{
X<-matrix(c(6,-4,1, -4,6,-4,1,-4,6), nrow=3,ncol=3, byrow = TRUE)
covmat <- matrix(rnorm(3 * 3), 3, 3)
H<-hatMatrix(X)
H<-hatMatrix(X, covmat)
diag(H)
}

