% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_optR.R
\name{inv.optR}
\alias{inv.optR}
\title{Invert a matrix using LU decomposition}
\usage{
inv.optR(A, tol = 1e-07)
}
\arguments{
\item{A}{: Input matrix}

\item{tol}{: tolerance}
}
\value{
A  : Inverse of Matrix A
}
\description{
function invert a matrix A using LU decomposition such that A*inv(A)=I
}
\examples{
# Invert the matrix using LU decomposition
A<-matrix(c(0.6,-0.4,1, -0.3,0.2,0.5,0.6,-1,0.5), nrow=3,ncol=3, byrow = TRUE) 
InvA<-inv.optR(A)
}

