% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtonRapshon.R
\name{jacobian}
\alias{jacobian}
\title{Function to evaluate jacobian matrix from functions}
\usage{
jacobian(f, x)
}
\arguments{
\item{f}{: function to optimize}

\item{x}{: Initial Solution}
}
\value{
jacobiabMatrix: Jacobian matrix

f0: Intial solution
}
\description{
jacobian is function to determine the jacobian matrix for function f for input x
}

