% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LU.R
\name{LU.decompose}
\alias{LU.decompose}
\title{LU decomposition}
\usage{
LU.decompose(A, tol = 1e-07)
}
\arguments{
\item{A}{: Input Matrix}

\item{tol}{: tolerance}
}
\value{
A  : Transformed matrix with Upper part of the triangele is (U) and Lower part of the triangular is L (Diagnoal for the L matrix is 1)
}
\description{
The function decomposes matrix A into LU with L lower matrix and U as upper matrix
}
\examples{
A<-matrix(c(0,-1,1, -1,2,-1,2,-1,0), nrow=3,ncol=3, byrow = TRUE)
Z<-optR(A, tol=1e-7, method="LU")
}

