% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LU.R
\name{LU.optR}
\alias{LU.optR}
\title{Solving system of equations using LU decomposition}
\usage{
LU.optR(A, b, tol = 1e-07)
}
\arguments{
\item{A}{: Input Matrix}

\item{b}{: Response}

\item{tol}{: tolerance}
}
\value{
U  : Upper part of the triangele is (U) and Lower part of the triangular is L (Diagnoal for the L matrix is 1)

c  : Lc=b (where Ux=c)

beta : Estimates
}
\description{
The function solves Ax=b using LU decomposition (LUx=b). The function handles multple responses
}
\examples{
A<-matrix(c(0,-1,1, -1,2,-1,2,-1,0), nrow=3,ncol=3, byrow = TRUE)
b<-matrix(c(0,0, 1), nrow=3,ncol=1,byrow=TRUE)
Z<-optR(A, b, method="LU")
}

