% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LUsplit}
\alias{LUsplit}
\title{Function to extract Lower and Upper matrix from LU decomposition}
\usage{
LUsplit(A)
}
\arguments{
\item{A}{: Input matrix}
}
\value{
U  : upper triangular matrix

L  : Lower triangular matrix
}
\description{
function to extract Lower and Upper matrix from LU decomposition
}
\examples{
A<-matrix(c(0,-1,1, -1,2,-1,2,-1,0), nrow=3,ncol=3, byrow = TRUE)
Z<-optR(A, method="LU")
LUsplit(Z$U)
}

