% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtonRapshon.R
\name{newtonRapson}
\alias{newtonRapson}
\title{Function for Newton Rapson roots for given equations}
\usage{
newtonRapson(f, x, iteration = 30, tol = 1e-09)
}
\arguments{
\item{f}{: function to optimize}

\item{x}{: Initial Solution}

\item{iteration}{: Iterations}

\item{tol}{: Tolerance}
}
\value{
x    : optimal roots
}
\description{
newtonRapson function perform optimization
}

