% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nonDiagMultipication}
\alias{nonDiagMultipication}
\title{Non-diagnoal multipication}
\usage{
nonDiagMultipication(i, A, beta)
}
\arguments{
\item{i}{: Column Index of Matrix A}

\item{A}{: Input matrix}

\item{beta}{: Response}
}
\value{
asum: Non-diagnol contribution
}
\description{
Function for non-diagnoal multipication
}

