% Generated by roxygen2 (4.0.1): do not edit by hand
\name{opt.matrix.reorder}
\alias{opt.matrix.reorder}
\title{Function to Re-order the matrix to make dominant diagnals}
\usage{
opt.matrix.reorder(A, tol = 1e-16)
}
\arguments{
\item{A}{: Input Matrix}

\item{tol}{: tolerance}
}
\value{
A        : Updated Matrix

b.order  : Order sequence of A updated  matrix
}
\description{
Function re-order the matrix to make matrix pivot.diag at each iteration
}
\examples{
A<-matrix(c(0,-1,1, -1,2,-1,2,-1,0), nrow=3,ncol=3, byrow = TRUE)
b<-matrix(c(0,0, 1), nrow=3,ncol=1,byrow=TRUE)
Z<-optR(A, b, method="gauss")
}

