% Generated by roxygen2 (4.0.1): do not edit by hand
\name{optR.backsubsitution}
\alias{optR.backsubsitution}
\title{Function to solve linear system using backsubsitution}
\usage{
\method{optR}{backsubsitution}(U, c)
}
\arguments{
\item{U}{: Upper triangular matrix}

\item{c}{: response}
}
\value{
beta : Estiamted value
}
\description{
Function to solve linear system using backsubsitution using Upper Triangular Matrix (Ux=c)
}

