% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optRFun.R
\name{optRFun.newtonRapson}
\alias{optRFun.newtonRapson}
\title{Function based optimization module}
\usage{
optRFun.newtonRapson(formula, x0, iteration = 30, tol = 1e-09)
}
\arguments{
\item{formula}{: Function to optimize}

\item{x0}{: Initial Solution}

\item{iteration}{: Number of Iterations}

\item{tol}{: Tolerance}
}
\value{
optRFun : Optimal Solution
}
\description{
Newton Rapshon based optimization
}

