% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optRFun.R
\name{optRFun}
\alias{optRFun}
\title{Function based optimization module}
\usage{
optRFun(formula, x0, iteration = 30, method = c("newtonrapson"),
  tol = 1e-09)
}
\arguments{
\item{formula}{: Function to optimize}

\item{x0}{: Initial Solution}

\item{iteration}{: Number of Iterations}

\item{method}{: Method for solving the optimization}

\item{tol}{: Tolerance}
}
\value{
optRFun : Optimal Solution class
}
\description{
Function based optimization module
}

