% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optR.R
\name{print.optR}
\alias{print.optR}
\title{print coefficients for optR class}
\usage{
\method{print}{optR}(x, ...)
}
\arguments{
\item{x}{: Input of optR class}

\item{...}{: S3 class}
}
\description{
optR is the default function for optimization
}
\examples{
# Solving equation Ax=b
A<-matrix(c(6,-4,1, -4,6,-4,1,-4,6), nrow=3,ncol=3, byrow = TRUE)
b<-matrix(c(-14,36, 6), nrow=3,ncol=1,byrow=TRUE)
Z<-optR(A, b, method="gauss")
print(Z)
}

