% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optR.R
\name{summary.optR}
\alias{summary.optR}
\title{Generate Summary for optR class}
\usage{
\method{summary}{optR}(object, ...)
}
\arguments{
\item{object}{: Input of optR class}

\item{...}{: S3 method}
}
\description{
summary function generates the summary for the optR class
}
\examples{
# Solving equation Ax=b
A<-matrix(c(6,-4,1, -4,6,-4,1,-4,6), nrow=3,ncol=3, byrow = TRUE)
b<-matrix(c(-14,36, 6), nrow=3,ncol=1,byrow=TRUE)
Z<-optR(A, b, method="cgm")
summary(Z)
}

