% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.R
\name{optR.multiplyfactor}
\alias{optR.multiplyfactor}
\title{Function to estimate lambda}
\usage{
\method{optR}{multiplyfactor}(rowindex, A, pivotindex)
}
\arguments{
\item{rowindex}{: Row Index for the row to be used}

\item{A}{: Input matrix}

\item{pivotindex}{: Column index for the pivot}
}
\value{
lambda     : Lambda
}
\description{
Function esimates the lambda or multiplier factor for Elimination using the pivot row/column
}

